requirements.txt
aiogram==3.0.0b7
SQLAlchemy==1.4.46
asyncpg==0.27.0
aiohttp==3.8.4
APScheduler==3.10.4
python-dotenv==1.0.0
psycopg2-binary==2.9.7